[REL-829] release-cli: get NNS proposal info with Rust+candid instead of using dfx

+++ migrations/2021-10-05-201137_subnet_update_nodes/up.sql
@@ -4,6 +4,6 @@ CREATE TABLE subnet_update_nodes (
   in_progress BOOLEAN NOT NULL DEFAULT true,
   nodes_to_add TEXT,
   nodes_to_remove TEXT,
-  proposal_add_id INTEGER,
-  proposal_remove_id INTEGER
+  proposal_add_id BIGINT,
+  proposal_remove_id BIGINT
 );
